refactor: create job route
