+++ prisma/schema.prisma
@@ -28,20 +28,19 @@ model Category {
 }
 
 model Job {
-  id             String          @id() @db.Uuid
+  id             String          @id() @db.Uuid @default(uuid())
   title          String          @db.VarChar
   description    String          @db.VarChar
   deliveryDate   DateTime        @db.Timestamptz(6) @map("delivery_date")
   cep            String          @db.VarChar
 
-  typeId         Int?            @map("type_id")
+  status         jobStatus?      @default(available)
   userId         String?         @db.Uuid @map("user_id")
   reviewId       String?         @unique() @db.Uuid @map("review_id")
   categoryId     Int?            @map("category_id")
 
   categories     Category?      @relation(fields: [categoryId], references: [id], onDelete: NoAction, onUpdate: NoAction)
   reviews        Review?        @relation(fields: [reviewId], references: [id], onDelete: NoAction, onUpdate: NoAction)
-  types_jobs     TypeJob?       @relation(fields: [typeId], references: [id], onDelete: NoAction, onUpdate: NoAction)
   users          User?          @relation(fields: [userId], references: [id], onDelete: NoAction, onUpdate: NoAction)
   candidates     Candidate?
   supplier_taken SupplierTaken?
@@ -49,8 +48,17 @@ model Job {
   @@map("jobs")
 }
 
+enum jobStatus{
+  available
+  doing
+  finished
+}
+
+
+
+
 model Review {
-  id      String  @id() @db.Uuid
+  id      String  @id() @db.Uuid  @default(uuid())
   score   Int
   comment String? @db.VarChar
   jobs    Job?
@@ -69,13 +77,7 @@ model SupplierTaken {
   @@map("supplier_taken")
 }
 
-model TypeJob {
-  id   Int    @id() @default(autoincrement())
-  name String @db.VarChar
-  jobs Job[]
 
-  @@map("types_jobs")
-}
 
 model User {
   id             String           @id() @default(uuid()) @db.Uuid
