+++ src/controllers/jobs/jobsCreate.controller.ts
@@ -2,16 +2,17 @@ import { Request, Response } from "express";
 import jobsCreateService from "../../services/jobs/createJob.servive";
 
 const jobsCreateController = async (request: Request, response: Response) => {
-  const { title, description, deliveryDate, cep, client, category } =
-    request.body;
+  const { title, description, deliveryDate, cep, category } = request.body;
+
+  const { userId } = request;
 
   const newJob = await jobsCreateService({
     title,
     description,
     deliveryDate,
     cep,
-    client,
-    category,
+    userId,
+    categoryId: category,
   });
 
   return response.status(201).json({
